function eternus_dx hard coded use of snmp
