+++ install/requirements.txt
@@ -6,4 +6,4 @@ requests
 python-dateutil
 grafanalib
 influxdb_client
-pysnmp
+easysnmp
